Test Windows run

+++ pandana/tests/test_cyaccess.py
@@ -33,9 +33,9 @@ def net(nodes_and_edges):
     edges["to"] = node_locations.loc[edges["to"]].values
 
     net = cyaccess(
-        nodes.index.values.astype('int32'),
+        nodes.index.values.astype('int64'),
         nodes.values,
-        edges.values.astype('int32'),
+        edges.values.astype('int64'),
         edge_weights.transpose().values,
         True
     )
